+++ src/main/java/core/aop/Hello.java
@@ -4,4 +4,6 @@ public interface Hello {
     String sayHello(String name);
     String sayHi(String name);
     String sayThankYou(String name);
+
+    String pinpong(String name);
 }
