+++ src/main/java/core/aop/HelloTarget.java
@@ -16,4 +16,9 @@ public class HelloTarget implements Hello{
     public String sayThankYou(String name) {
         return "Thank You " + name;
     }
+
+    @Override
+    public String pinpong(String name) {
+        return "ping pong " + name;
+    }
 }
